+++ src/elements/conditions/podcasts/PodcastIsNewFeedUrlRule.php
@@ -47,6 +47,6 @@ class PodcastIsNewFeedUrlRule extends BaseLightswitchConditionRule implements El
     public function matchElement(ElementInterface $element): bool
     {
         /** @var Podcast $element */
-        return $this->matchValue($element->podcastIsNewFeedUrl !== null);
+        return $this->matchValue($element->podcastIsNewFeedUrl);
     }
 }